fix warnings

+++ Source/LuminoCore/Source/Base/ByteBuffer.cpp
@@ -256,10 +256,8 @@ int ByteBuffer::compare(const ByteBuffer& buf1, const void* buf2, size_t buf2Siz
 	// TODO: いろいろだめ。
 	if (count >= 0)
 	{
-		if (buf1.getSize() < count) return -1;
-		if (buf2Size < count) return 1;
-		//if (buf1.GetSize() > count) return 1;
-		//if (buf1.GetSize() == 0) return 0;			// 両方 0
+		if (buf1.getSize() < (size_t)count) return -1;
+		if (buf2Size < (size_t)count) return 1;
 	}
 	else
 	{
